Change Preview size variable to size_t

+++ src/preview.cpp
@@ -1098,7 +1098,7 @@ namespace Exiv2 {
         return pData_;
     }
 
-    uint32_t PreviewImage::size() const
+    size_t PreviewImage::size() const
     {
         return size_;
     }
